+++ src/chameleon/zpt/template.py
@@ -1,6 +1,7 @@
 from functools import partial
 from hashlib import sha256
 from os.path import dirname
+from zipfile import Path
 
 from ..astutil import Symbol
 from ..compiler import ExpressionEngine
@@ -256,7 +257,7 @@ class PageTemplate(BaseTemplate):
             body = body.replace('\r\n', '\n').replace('\r', '\n')
 
         return MacroProgram(
-            body, self.mode, self.spec,
+            body, self.mode, self.filename,
             escape=True if self.mode == "xml" else False,
             default_marker=self.default_marker,
             boolean_attributes=boolean_attributes or frozenset([]),
@@ -408,7 +409,7 @@ class PageTemplateFile(PageTemplate, BaseTemplateFile):
 
     prepend_relative_search_path = True
 
-    def __init__(self, spec, search_path=None, loader_class=TemplateLoader,
+    def __init__(self, filename, search_path=None, loader_class=TemplateLoader,
                  **config):
         if search_path is None:
             search_path = []
@@ -422,7 +423,10 @@ class PageTemplateFile(PageTemplate, BaseTemplateFile):
             # If the flag is set (this is the default), prepend the path
             # relative to the template file to the search path
             if self.prepend_relative_search_path:
-                path = dirname(self.spec.filename)
+                if isinstance(self.filename, Path):
+                    path = self.filename.parent
+                else:
+                    path = dirname(self.filename)
                 search_path.insert(0, path)
 
             loader = loader_class(search_path=search_path, **config)
@@ -433,7 +437,7 @@ class PageTemplateFile(PageTemplate, BaseTemplateFile):
             self._loader = loader.bind(template_class)
 
         super().__init__(
-            spec,
+            filename,
             post_init_hook=post_init,
             **config
         )